fix: switch to panic (exception) over fatal (interrupt)
Use panic to allow scenarios to handle exeptions and prevent other
K6 iterations for the same VU from being impacted.

To properly be considered an exception by goja, we need to include
a `Value`. The Runtime contains a ToValue that we can use to convert
go strings to a Value accesible in JS.

The runtime is only accessible on the module instance's VU. This
requires some refactoring to the module init and wrapper structs